requirements.txt
annotated-types==0.6.0
anyio==4.2.0
certifi==2024.2.2
cffi==1.16.0
click==8.1.7
coverage==7.3.2
cryptography==42.0.2
dnspython==2.5.0
email-validator==2.1.0.post1
exceptiongroup==1.2.0
fastapi==0.109.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
iniconfig==2.0.0
jwt==1.3.1
motor==3.3.2
mysql-connector-python==8.3.0
packaging==23.2
passlib==1.7.4
pluggy==1.4.0
prettytable==3.9.0
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
PyJWT==2.8.0
pymongo==4.6.1
pytest==8.0.0
python-dotenv==1.0.1
sniffio==1.3.0
starlette==0.35.1
tabulate==0.9.0
TatSu==5.8.3
tomli==2.0.1
typing_extensions==4.9.0
uvicorn==0.27.0
wcwidth==0.2.13
websockets==12.0
